Added user DTOs and validations

+++ GonoPic/Controllers/UserController.cs
@@ -1,5 +1,6 @@
-﻿using GonoPic.Business.Services.Interfaces;
-using GonoPic.Data.Entities;
+﻿using GonoPic.Business.DTOs;
+using GonoPic.Business.Mappers;
+using GonoPic.Business.Services.Interfaces;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -17,31 +18,42 @@ namespace GonoPic.Controllers
         }
 
         [HttpGet]
-        public async Task<ActionResult<IEnumerable<User>>> GetAll()
+        public async Task<ActionResult<IEnumerable<UserDto>>> GetAll()
         {
             var users = await _userService.GetAllUsers();
-            return Ok(users);
+            var userDtos = users.Select(UserMapper.ToDto);
+            return Ok(userDtos);
         }
 
         [HttpGet("{id}")]
-        public async Task<ActionResult<User>> Get(int id)
+        public async Task<ActionResult<UserDto>> Get(int id)
         {
             var user = await _userService.GetUserById(id);
-            if (user == null) return NotFound();
-            return Ok(user);
+            if (user == null) 
+                return NotFound();
+
+            var userDto = UserMapper.ToDto(user);
+            return Ok(userDto);
         }
 
         [HttpPost]
-        public async Task<ActionResult> Create(User user)
+        public async Task<ActionResult> Create(UserCreateDto dto)
         {
+            var user = UserMapper.ToEntity(dto);
             await _userService.CreateUser(user);
-            return CreatedAtAction(nameof(Get), new { id = user.Id }, user);
+            
+            var userDto = UserMapper.ToDto(user);
+            return CreatedAtAction(nameof(Get), new { id = user.Id }, userDto);
         }
 
         [HttpPost("{id}")]
-        public async Task<ActionResult> Update(int id, User user)
+        public async Task<ActionResult> Update(int id, UserUpdateDto dto)
         {
-            if (id != user.Id) return BadRequest();
+            var user = await _userService.GetUserById(id);
+            if (user == null) 
+                return NotFound();
+
+            UserMapper.UpdateEntity(user, dto);
             await _userService.UpdateUser(user);
             return NoContent();
         }
@@ -49,6 +61,10 @@ namespace GonoPic.Controllers
         [HttpDelete("{id}")]
         public async Task<ActionResult> Delete(int id)
         {
+            var user = await _userService.GetUserById(id);
+            if (user == null)
+                return NotFound();
+
             await _userService.DeleteUser(id);
             return NoContent();
         }
